+++ NEWS.md
@@ -1,3 +1,10 @@
 ## CytoPipeline 0.99.0
 
 - Prior to Bioconductor submission
+
+## CytoPipeline 0.99.1
+
+- Maintenance due to Bioc version change (3.17)
+- removed use of openCyto::gate_tail() (disappeared w/o deprecation), replaced
+by flowDensity::deGate()
+